mean_table/requirements.txt
paho-mqtt
pandas
rich

